Local file restrictions should not block sessionStorage access. https://bugs.webkit.org/show_bug.cgi?id=155609 <rdar://problem/25229461>
Reviewed by Andy Estes.

Re-landing this fix that was lost when the localStorage change was rolled out.

Use of 'sessionStorage' is governed by SecurityOrigin with third party access
set to 'ShouldAllowFromThirdParty::AlwaysAllowFromThirdParty'. We should not
reject local files for this combination of arguments.

Tested by storage/domstorage/sessionstorage/blocked-file-access.html.

* page/SecurityOrigin.cpp:
(WebCore::SecurityOrigin::canAccessStorage): For the case of sessionStorage,
allow local file access.



Identifier: 182273@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@208550 268f45cc-cd09-0410-ab3c-d52691b4dbfc