[CSS Parser] Fix errors in numerous layout tests https://bugs.webkit.org/show_bug.cgi?id=165343
Reviewed by Zalan Bujtas.

* fast/backgrounds/repeat/parsing-background-repeat-expected.txt:
* fast/backgrounds/repeat/script-tests/parsing-background-repeat.js:
The old parser partially applies shorthands, even if it rejects later on.
The new parser properly rejects the entire shorthand. Fix the test so that
the numbers at the end are not unitless so that the entire shorthand parses, allowing
the components to be obtained.

* fast/css/unknown-pseudo-element-matching-expected.txt:
* fast/css/unknown-pseudo-element-matching.html:
* fast/text/crash-complex-text-surrogate.html:
The old parser allows the tag name to be omitted following a namespace bar (|). This
is not legal according to the namespaces spec, and the new parser behaves correctly.
For the crash test, it's important that all the code still runs, so we know the crash
doesn't happen, so I patched the test to put in '*' for the tag names to keep everything
the same.

For the unknown-pseudo-element test, I just removed the five invalid cases.



Identifier: 182956@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@209279 268f45cc-cd09-0410-ab3c-d52691b4dbfc