Two tiled drawing tests failing with visual viewports enabled. https://bugs.webkit.org/show_bug.cgi?id=165489
Reviewed by Dean Jackson.
Source/WebCore:

computeLayoutViewportOrigin() gets called for iframes when the layout viewport
is zero-sized, but the visual viewport is non-zero. It doesn't really make sense
to compute a layout viewport when the visual viewport is larger than the layout
viewport, but if this happens just anchor the layout viewport at the origin
of the visual viewport.

* page/FrameView.cpp:
(WebCore::FrameView::computeLayoutViewportOrigin):

LayoutTests:

Unfail and rebaseline.

* platform/mac-wk2/TestExpectations:
* tiled-drawing/scrolling/frames/coordinated-frame-expected.txt:
* tiled-drawing/scrolling/frames/coordinated-frame-in-fixed-expected.txt:

Identifier: 183114@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@209447 268f45cc-cd09-0410-ab3c-d52691b4dbfc